// src/newFontTemplate.generated.ts
export type TemplateGlyph = { name: string; unicode?: string };